refactor: move ExecuteCommandTool from distri lib to distri-cli
ExecuteCommandTool is CLI-specific (local shell execution), not a
shared library concern. Move it to distri-cli/src/tools.rs.

distri (shared lib) keeps register_local_filesystem_tools — the
generic mechanism any client uses for filesystem tool registration.
